# docs/swagger.yaml
definitions:
  api.Response:
    properties:
      code:
        description: 状态码
        example: 200
        type: integer
      data:
        description: 数据
      message:
        description: 消息
        example: Success
        type: string
      status:
        description: 状态
        type: boolean
    type: object
...
  form.UserAdminUpdateForm:
    properties:
      admin:
        type: boolean
      password:
        type: string
    type: object
  form.UserConfigUpdateForm:
    properties:
      autoCenter:
        type: boolean
      iconSize:
        type: integer
      lat:
        type: number
      link:
        type: string
      lon:
        type: number
      maxZoom:
        type: integer
      minZoom:
        type: integer
      note:
        type: string
      title:
        type: string
      tolorance:
        type: number
      zoom:
        type: integer
    type: object
  form.UserCreateForm:
    properties:
      admin:
        type: boolean
      password:
        minLength: 6
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  form.UserLoginForm:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8090
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /api/v1/share/{username}:
    get:
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      - description: group name
        in: query
        name: group
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/api.Response'
      summary: user share place
      tags:
      - user
...
  /api/v1/user/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/form.UserCreateForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - JWT: []
      summary: create user
      tags:
      - user
  /api/v1/user/config:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - JWT: []
      summary: user config
      tags:
      - user
  /api/v1/user/config/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/form.UserConfigUpdateForm'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - JWT: []
      summary: update user map config
      tags:
      - user
  /api/v1/user/del/{id}:
    get:
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - JWT: []
      summary: delete user
      tags:
      - user
  /api/v1/user/login:
    post:
      parameters:
      - description: UserLoginForm
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/form.UserLoginForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      summary: user login
      tags:
      - user
  /api/v1/user/logout:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - JWT: []
      summary: user login
      tags:
      - user
  /api/v1/user/update/{id}:
    post:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/form.UserAdminUpdateForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - JWT: []
      summary: update user
      tags:
      - user
  /api/v1/users:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - JWT: []
      summary: get all users
      tags:
      - user
securityDefinitions:
  JWT:
    description: |-
      Bearer Token 格式：`Bearer {token}`
      JWT Authorization header
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"